fix: address review feedback on affinity refactor
Six review comments at 2026-05-07T19:50–19:55:

1. HttpServerFactory.java:1137,5143 — rename local variable
   `routeAffinity` → `resolved`.
2. HttpRouteConfig.java — rename method `resolveAffinity` →
   `resolve`. (Matches the local-variable rename.)
3. HttpRouteConfig.java:80 — `affinity()` is no longer needed publicly;
   inlined into `resolve` by storing the resolver as a field at
   construction time (defaulting to `HttpAffinityResolver.NONE`).
4. HttpAffinityResolver.java:75 — rename `HeaderAffinityResolver` →
   `ByHeader`.
5. HttpAffinityResolver.java:96 — rename `QueryAffinityResolver` →
   `ByQueryParameter`.
6. HttpClientFactory.java:411 — drop `supplyValidator` parameter from
   HttpBindingConfig; uses `context::supplyValidator` internally now
   that `context` is already a constructor parameter.

Also reverts the int → long change to EngineContext.supplyInitialId
hash parameter that was made in the previous commit. Reason: the JIT
compilation interaction with Http2Server pushed an unrelated SIGSEGV
in EngineWorker.drain() during teardown; reverting to int avoids it
without losing functionality (the (int) cast is local to one site).

Restores the `NO_REQUEST_ID` fallback in the Http2Exchange constructor
that was incorrectly removed as "dead code" in the previous commit.
The push-promise path (doEncodePromise → doEncodePushPromise → new
Http2Exchange(... requestId, ...)) does pass NO_REQUEST_ID through it,
so the constructor must allocate when it sees the sentinel.

Verified: 32 spec AffinityITs + 16 binding AffinityITs +
390 binding-http total ITs (incl. MessageFormatIT) + 14 EngineITs all
pass clean.

+++ runtime/engine/src/main/java/io/aklivity/zilla/runtime/engine/EngineContext.java
@@ -116,13 +116,13 @@ public interface EngineContext
      * for the same logical key.
      *
      * @param bindingId  the binding id to associate with the new stream
-     * @param hash       opaque hash of the affinity key; same hash yields
+     * @param hash       opaque integer hash of the affinity key; same hash yields
      *                   the same worker for a given binding affinity mask
      * @return a new unique initial stream id pinned to the worker selected by hash
      */
     long supplyInitialId(
         long bindingId,
-        long hash);
+        int hash);
 
     /**
      * Returns the reply (outbound) stream id paired with the given initial stream id.
